pyproject.toml
[project]
name = "chroma-photo-search"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chromadb>=1.0.9",
    "matplotlib>=3.10.3",
    "numpy>=2.2.6",
    "open-clip-torch>=2.32.0",
    "openai>=1.81.0",
    "pillow>=11.2.1",
    "tqdm>=4.67.1",
]
